backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
chromadb==0.5.5
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-ollama
llama-index-llms-ollama
python-dotenv==1.0.1
